Update navigational bar to make it less crowded
Shorten navigational items and reorder them a bit to make the
navigational bar a little more accessible and less crowded. Also
put "Download data" last in the site navigation instead of having
it somewhere in the middle because it serves a different purpose
than the other items in the navbar.